requirements.txt
-i https://pypi.org/simple
aenum==3.1.12
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0
requests==2.30.0
requests-toolbelt==1.0.0
six==1.16.0
typing-extensions==4.5.0
urllib3==2.0.2
wxc-sdk==1.14.0
yarl==1.9.2
